refactor(query): 调整用户行为与视频系列查询逻辑

+++ only-danmuku-adapter/src/main/kotlin/edu/only4/danmuku/adapter/portal/api/payload/AccountUserCountInfo.kt
@@ -3,7 +3,7 @@ package edu.only4.danmuku.adapter.portal.api.payload
 object AccountUserCountInfo {
     data class Response(
         val fansCount: Long,
-        val currentCoinCount: Long,
+        val currentCoinCount: Int,
         val focusCount: Long
     )
 }
